requirements.txt
Flask==2.3.3
python-dotenv==1.1.0
Pillow==11.1.0
transformers>=4.45.2
accelerate>=0.20.0
torch>=2.7.1